backend/app/requirements.txt
Flask==0.12.1
redis==2.10.5
ruamel.yaml==0.14.5
uWSGI==2.0.15
